fixed receipt

+++ q_breaker_app/q_breaker_app/lib/fetching_data.dart
@@ -232,7 +232,7 @@ class Food{
 
   double getNetCost(){
     netCost = quantity * price;
-    return netCost;
+    return netCost.toDouble();
   }
 
   void changeReadyState(){
